java_testgen_app/src/main/java/ChatGPT/Dijkstra/AIGenerated/T2/Dijkstra.java
package ChatGPT.Dijkstra.AIGenerated.T2;

import java.util.*;

public class Dijkstra {
    private List<Edge> edges;
    private int n;

    public Dijkstra(List<Edge> edges, int n) {
        this.edges = edges;
        this.n = n;
    }

    public Map<Integer, Integer> findShortestPaths(int source) {
        Map<Integer, Integer> shortestPaths = new HashMap<>();
        PriorityQueue<Node> minHeap = new PriorityQueue<>(Comparator.comparingInt(node -> node.weight));
        boolean[] visited = new boolean[n];

        // Initialize distances to infinity and mark all nodes as unvisited
        int[] distances = new int[n];
        Arrays.fill(distances, Integer.MAX_VALUE);

        distances[source] = 0;
        minHeap.offer(new Node(source, 0));

        while (!minHeap.isEmpty()) {
            Node currentNode = minHeap.poll();
            int currentVertex = currentNode.vertex;

            if (visited[currentVertex]) {
                continue;
            }

            visited[currentVertex] = true;

            for (Edge edge : edges) {
                if (edge.source == currentVertex) {
                    int neighborVertex = edge.dest;
                    int totalWeight = distances[currentVertex] + edge.weight;

                    if (totalWeight < distances[neighborVertex]) {
                        distances[neighborVertex] = totalWeight;
                        minHeap.offer(new Node(neighborVertex, totalWeight));
                    }
                }
            }
        }

        // Populate the shortestPaths map with the computed distances
        for (int i = 0; i < n; i++) {
            if (distances[i] != Integer.MAX_VALUE) {
                shortestPaths.put(i, distances[i]);
            }
        }

        return shortestPaths;
    }

    public static class Edge {
        public int source;
        public int dest;
        public int weight;

        public Edge(int source, int dest, int weight) {
            this.source = source;
            this.dest = dest;
            this.weight = weight;
        }
    }

    public static class Node {
        public int vertex;
        public int weight;

        public Node(int vertex, int weight) {
            this.vertex = vertex;
            this.weight = weight;
        }
    }
}
